Cancel spot requests when associated instances are terminated
When using one-time spot requests, the instances are terminated, but the spot request is not cancelled, leaving the door open to being filled again in the future. This PR deletes the spot request if it exists in the aws delete_proxy() function. Related to #42, but unclear if applicable to persistent spot requests.